Remove unused comments after function move
