+++ src/inputHandlers/number.js
@@ -1,5 +1,33 @@
-import { ensureNumberInput } from '../browser/toys.js';
+import {
+  createNumberInput,
+  createUpdateTextInputValue,
+} from '../browser/toys.js';
 import { maybeRemoveElement } from './disposeHelpers.js';
+
+const positionNumberInput = ({ container, textInput, numberInput, dom }) => {
+  const nextSibling = dom.getNextSibling(textInput);
+  container.insertBefore(numberInput, nextSibling);
+};
+
+export const ensureNumberInput = (container, textInput, dom) => {
+  let numberInput = dom.querySelector(container, 'input[type="number"]');
+
+  if (!numberInput) {
+    numberInput = createNumberInput(
+      textInput.value,
+      createUpdateTextInputValue(textInput, dom),
+      dom
+    );
+    positionNumberInput({
+      container,
+      textInput,
+      numberInput,
+      dom,
+    });
+  }
+
+  return numberInput;
+};
 
 function maybeRemoveKV(container, dom) {
   const kvContainer = dom.querySelector(container, '.kv-container');
